requirements.txt
fastapi
uvicorn[standard]
python-dotenv

pydantic

langchain
langchain_core
langchain-google-genai
langgraph

langsmith

llama-cloud

chromadb

nest-asyncio
opentelemetry-api
opentelemetry-sdk

python-multipart